langgraph-supervisor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.0,<3.0.0
python-dotenv==1.2.1
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=1.0.1
httpx>=0.25.2
langchain-core>=0.1.0
